Added Current Sensor and Filters

+++ src/IVoltageSensor.h
@@ -3,7 +3,7 @@
 
 class IVoltageSensor {
     public:
-        virtual float getVoltage() = 0;
+        virtual double getVoltage() = 0;
 };
 
 #endif
